refactor(interfaces/chatbot): spliting indext into other functions
