+++ src/interfaces/chatbot/index.ts
@@ -1,28 +1,11 @@
-import TwitchJs from "twitch-js";
-import registerListeners from './listeners'
+import { chat, channel, stop } from './chat'
+import start from './start'
 
-const {
-    TWITCH_BOT_TOKEN,
-    TWITCH_CHANNEL,
-} = process.env
-
-const twitchAuth = {
-    username: 'victor_perin_bot',
-    token: TWITCH_BOT_TOKEN,
-    log: { level: 'error' },
-}
-
-export const { chat } = new TwitchJs(twitchAuth)
-export const channel = TWITCH_CHANNEL
-
-
-export const start = async (): Promise<void> => {
-    await registerListeners()
-    await chat.connect()
-    await chat.join(channel)
-}
-
-export const stop = (): void => {
-    chat.removeAllListeners()
-    chat.disconnect()
+export { chat, channel, stop } from './chat'
+export { default as start } from './start'
+export default {
+    chat,
+    channel,
+    stop,
+    start,
 }
