+++ src/interfaces/chatbot/start.ts
@@ -1,0 +1,8 @@
+import { chat, channel } from './chat'
+import registerListeners from './listeners'
+
+export default async (): Promise<void> => {
+    await registerListeners()
+    await chat.connect()
+    await chat.join(channel)
+}
